Nics: Managed private network API change
The `listNics` call will return IP addresses associated to nics in
managed private networks even if the IP is not static.

Terraform should only update the IP address associated to a nic if an
IP is set in the Terraform file, and if the IP returned by `listNics`
is different.