compute: replace template attribute with template_id
This change has been prompted by issue #8, which illustrates that it's
not possible to reference an `exoscale_compute` resource `template` by
an ID (only a name) -- effectively preventing the usage of custom
templates via the `exoscale_compute_template` data source.

As a consequence, the `template` attribute is now deprecated and
replaced by a new `template_id` attribute which only accepts template
IDs, which can be looked up via the `exoscale_compute_template` DS.
A collateral damage of this change is that the `username` attribute of
the resource is now also deprecated and replaced by the
`exoscale_compute_template` DS `username` attribute.

Fixes #8.