compute: Recompute `ip_address` attribute on state update
This commit introduces custom diff logic to ensure a compute resource's
`ip_address` attribute is correctly recomputed if the old value of the
attribute is `""`. This transition occurs if the resource is initially
provisioned in "Stopped" state and then booted for the first time in a
subsequent `terraform apply` step.

Note that the custom diff logic assumes that a VM's IP address remains
stable over its life time, and  only forces recomputation of the
`ip_address` attribute when the old IP address attribute is the empty
string and the VM's state changes from "Stopped" to "Running". Doing so
minimizes the churn on resources which depend on the VM's IP address
attribute such as DNS records.

The implementation makes use of the `helper/customdiff` package of the
Terraform plugin SDK to implement the diffing logic.

See also:
* https://www.terraform.io/docs/extend/resources/customizing-differences.html
* https://pkg.go.dev/github.com/hashicorp/terraform-plugin-sdk/helper/customdiff#section-documentation