Don't throw error if volume detach call fails due to volume not attached (#367)
This PR fixes the race condition that may trigger error when both volume
and volume attached instance are destroyed at the same time. As instance
deletion automatically detaches volume, API throws error when detach is
called as part of volume destroy.

Originally there was a check for volume state before detach is called,
but that has proven unreliable. Here we are executing detach, but
ignoring error if it says that volume is not attached.