Examples : add csi parameter in the example for SKS (#413)
# Description
This PR follows a customer feedback. He used the example to create a SKS
Cluster and then struggled with the PVC.
This, because the CSI addon is **NOT** deployed by default. 

Note : metrics_server and exoscale_ccm both default to true. 

As the auto_upgrade should not break, I suggest to set the boolean to
true.

<!--
* Prefix: the title with the component name being changed. Add a short
and self describing sentence to ease the review
* Please add a few lines providing context and describing the change
* Please self comment changes whenever applicable to help with the
review process
* Please keep the checklist as part of the PR. Tick what applies to this
change.
-->

## Checklist
(For exoscale contributors)

* [ ] Changelog updated (under *Unreleased* block)
* [ ] Acceptance tests OK
* [ ] For a new resource, datasource or new attributes: acceptance test
added/updated

## Testing

<!--
Describe the tests you did
-->