dbaas support for mysql,pg database (#410)
# Description
<!--
* Prefix: the title with the component name being changed. Add a short
and self describing sentence to ease the review
* Please add a few lines providing context and describing the change
* Please self comment changes whenever applicable to help with the
review process
* Please keep the checklist as part of the PR. Tick what applies to this
change.
-->

New resource introduced:
- `exoscale_dbaas_pg_database`, `exoscale_dbaas_mysql_database`, to
represent individual databases in database services.

To avoid confusion with the existing `exoscale_database` resource, it is
being *deprecated*. A new resource is introduced, `exoscale_dbaas`. The
only thing that changes is the name, under the hood those resources are
exactly the same, except a deprecation message for the former.

(I couldn't find guidance on how to rename a resource type so I have no
idea if it's possible to automate the move for users)

## Checklist
(For exoscale contributors)

* [x] Changelog updated (under *Unreleased* block)
* [x] Acceptance tests OK
* [x] For a new resource, datasource or new attributes: acceptance test
added/updated

## Testing

<!--
Describe the tests you did
-->

Acceptance tests updated